requirements.txt
flask
chatterbox-streaming
torch
numpy scipy